Mass source download fails on zero record members #2513
When trying to copy an empty member to an already existing member
the system issues CPF2869.

When the empty member is copied to a new member, the message is
not send and the transaction continues.

The fix is to use the member name for the temporary member
and not reuse the existing 'TEMPMEMBER' member.

This also fixes:
Strange dialog button when downloading members #2527